fix: listen for zoom on all layers
